pyproject.toml
[project]
name = "document-toword"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "reportlab>=4.4.5",
    "requests>=2.32.5",
    "xlrd>=2.0.2",
    "xlwt>=1.3.0",
]
